+++ src/TestHarness/Data/TestAPI.cs
@@ -1,22 +1,20 @@
 ﻿using Common.Core.Interfaces;
 using Common.Core.Models;
-using Common.Data.Sql;
+using Common.Data.Api;
 
-namespace TestHarness;
+namespace TestHarness.Data;
 
-internal class TestSql
+internal class TestAPI
 {
-	internal const string cTestDataDir = @"C:\Temp\TestData";
-
 	internal static bool RunTest()
 	{
-		_ = Program.sLogger.Info( "Testing Common.Data.Sql (Microsoft.Data.SqkClient)" );
-		_ = Program.sLogger.Info( "AddressData, CompanyData, and PersonData database connection strings" );
-		_ = Program.sLogger.Info( "Uses SQL Server Express databases (AddressData.mdf and EntityData.mdf)" );
+		_ = Program.sLogger.Info( "Testing Common.Data.Api (EFCore.RestApi)" );
+		_ = Program.sLogger.Info( "CommonData endpoint connection string" );
+		_ = Program.sLogger.Info( "Uses EFCommonData database with Common user login." );
 		Console.WriteLine();
 
 		// Test the address data
-		_ = new AddressData( useAlpha2: false );
+		_ = new AddressData( Program.sApp.ConfigFile, useAlpha2: false );
 		AddrParams args = new();
 		foreach( string code in TestAddress.postcodes )
 		{
@@ -33,7 +31,7 @@ internal class TestSql
 
 		#region Test the Company data
 
-		Companies companies = new();
+		Companies companies = new( Program.sApp.ConfigFile );
 		_ = Program.sLogger.Info( $"Companies total.: {companies.TotalCount:00#}" );
 
 		// Get a list of 5 Companies
@@ -45,10 +43,6 @@ internal class TestSql
 		IList<ICompany> listc = companies.Get( 5 );
 		_ = Program.sLogger.Info( $"Companies list..: {listc.Count:00#}" );
 
-		//var data = companies.Get( cTestDataDir, "Company-test.json", max: 5 );
-		//var data = companies.Get( cTestDataDir, "Company-test1.json", max: 15 );
-		//_ = companies.Serialize( cTestDataDir, "Company-testout.json", data );
-
 		// Get a specific Company
 		ICompany? company = companies.Find( listc[0].Id );
 		_ = company is not null
@@ -59,7 +53,7 @@ internal class TestSql
 
 		#region Test the Person data
 
-		People people = new();
+		People people = new( Program.sApp.ConfigFile );
 		_ = Program.sLogger.Info( $"People total....: {people.TotalCount:00#}" );
 
 		// Get a list of 10 People
@@ -71,12 +65,8 @@ internal class TestSql
 		IList<IPerson> listp = people.Get( 10 );
 		_ = Program.sLogger.Info( $"People list.....: {listp.Count:00#}" );
 
-		//var data = people.Get( cTestDataDir, "Person-test.json", max: 5 );
-		//var data = people.Get( cTestDataDir, "Person-test1.json", max: 15 );
-		//_ = people.Serialize( cTestDataDir, "Person-testout.json", data );
-
 		// Get a specific Person
-		IPerson? person = people.Find( listp[0].Id );
+		IPerson? person = people.Find( listc[0].Id );
 		_ = person is not null
 			? Program.sLogger.Info( $"Person Id  {listp[0].Id:00#} is {person.FullName}" )
 			: Program.sLogger.Info( $"Person Id {listp[0].Id:00#} not found!" );
