+++ src/TestHarness/Data/TestSql.cs
@@ -1,16 +1,18 @@
 ﻿using Common.Core.Interfaces;
 using Common.Core.Models;
-using Common.Data.SQLite;
+using Common.Data.Sql;
 
-namespace TestHarness;
+namespace TestHarness.Data;
 
-internal class TestSQLite
+internal class TestSql
 {
+	internal const string cTestDataDir = @"C:\Temp\TestData";
+
 	internal static bool RunTest()
 	{
-		_ = Program.sLogger.Info( "Testing Common.Data.SQLite (EFCore.Database.Entity)" );
-		_ = Program.sLogger.Info( "AddressDb and EntityDb database connection strings" );
-		_ = Program.sLogger.Info( "Uses SQLite databases (AddressData.db and EntityData.db)" );
+		_ = Program.sLogger.Info( "Testing Common.Data.Sql (Microsoft.Data.SqkClient)" );
+		_ = Program.sLogger.Info( "AddressData, CompanyData, and PersonData database connection strings" );
+		_ = Program.sLogger.Info( "Uses SQL Server Express databases (AddressData.mdf and EntityData.mdf)" );
 		Console.WriteLine();
 
 		// Test the address data
@@ -29,52 +31,55 @@ internal class TestSQLite
 		if( !TestAddress.RunTest( args ) ) { return false; }
 		Console.WriteLine();
 
-		EntityContextBase ctx = new();
-		// Connection string in debug:
-		// context => ChangeTracker => Context => Database => Non-Public members =>
-		//  Dependencies => RelationalConnection
-
 		#region Test the Company data
 
-		Companies companies = new( ctx );
+		Companies companies = new();
 		_ = Program.sLogger.Info( $"Companies total.: {companies.TotalCount:00#}" );
 
 		// Get a list of 5 Companies
 		if( companies.TotalCount < 5 )
 		{
-			_ = Program.sLogger.Info( "Companies count is less than 5" );
+			_ = Program.sLogger.Info( "Companies count is less than 5!" );
 			return false;
 		}
 		IList<ICompany> listc = companies.Get( 5 );
 		_ = Program.sLogger.Info( $"Companies list..: {listc.Count:00#}" );
 
+		//var data = companies.Get( cTestDataDir, "Company-test.json", max: 5 );
+		//var data = companies.Get( cTestDataDir, "Company-test1.json", max: 15 );
+		//_ = companies.Serialize( cTestDataDir, "Company-testout.json", data );
+
 		// Get a specific Company
 		ICompany? company = companies.Find( listc[0].Id );
 		_ = company is not null
 			? Program.sLogger.Info( $"Company Id {listc[0].Id:00#} is {company.Name}" )
-			: Program.sLogger.Info( $"Company Id {listc[0].Id:00#} not found." );
+			: Program.sLogger.Info( $"Company Id {listc[0].Id:00#} not found!" );
 
 		#endregion
 
 		#region Test the Person data
 
-		People people = new( ctx );
+		People people = new();
 		_ = Program.sLogger.Info( $"People total....: {people.TotalCount:00#}" );
 
 		// Get a list of 10 People
 		if( people.TotalCount < 10 )
 		{
-			_ = Program.sLogger.Info( "People count is less than 10" );
+			_ = Program.sLogger.Info( "People count is less than 10!" );
 			return false;
 		}
 		IList<IPerson> listp = people.Get( 10 );
 		_ = Program.sLogger.Info( $"People list.....: {listp.Count:00#}" );
 
+		//var data = people.Get( cTestDataDir, "Person-test.json", max: 5 );
+		//var data = people.Get( cTestDataDir, "Person-test1.json", max: 15 );
+		//_ = people.Serialize( cTestDataDir, "Person-testout.json", data );
+
 		// Get a specific Person
 		IPerson? person = people.Find( listp[0].Id );
 		_ = person is not null
-			? Program.sLogger.Info( $"Person Id. {listp[0].Id:00#} is {person.FullName}" )
-			: Program.sLogger.Info( $"Person Id. {listp[0].Id:00#} not found." );
+			? Program.sLogger.Info( $"Person Id  {listp[0].Id:00#} is {person.FullName}" )
+			: Program.sLogger.Info( $"Person Id {listp[0].Id:00#} not found!" );
 
 		#endregion
