+++ src/TestHarness/Data/TestSqlServer.cs
@@ -1,20 +1,20 @@
 ﻿using Common.Core.Interfaces;
 using Common.Core.Models;
-using Common.Data.Api;
+using Common.Data.SqlServer;
 
-namespace TestHarness;
+namespace TestHarness.Data;
 
-internal class TestAPI
+internal class TestSqlServer
 {
 	internal static bool RunTest()
 	{
-		_ = Program.sLogger.Info( "Testing Common.Data.Api (EFCore.RestApi)" );
-		_ = Program.sLogger.Info( "CommonData endpoint connection string" );
-		_ = Program.sLogger.Info( "Uses EFCommonData database with Common user login." );
+		_ = Program.sLogger.Info( "Testing Common.Data.SqlServer (EFCore.Database.Full)" );
+		_ = Program.sLogger.Info( "CommonData database connection string" );
+		_ = Program.sLogger.Info( "Uses SQL Server databases (TestEF)" );
 		Console.WriteLine();
 
 		// Test the address data
-		_ = new AddressData( Program.sApp.ConfigFile, useAlpha2: false );
+		_ = new AddressData( useAlpha2: false );
 		AddrParams args = new();
 		foreach( string code in TestAddress.postcodes )
 		{
@@ -29,15 +29,20 @@ internal class TestAPI
 		if( !TestAddress.RunTest( args ) ) { return false; }
 		Console.WriteLine();
 
+		FullContextBase ctx = new();
+		// Connection string in debug:
+		// context => ChangeTracker => Context => Database => Non-Public members =>
+		//  Dependencies => RelationalConnection
+
 		#region Test the Company data
 
-		Companies companies = new( Program.sApp.ConfigFile );
+		Companies companies = new( ctx );
 		_ = Program.sLogger.Info( $"Companies total.: {companies.TotalCount:00#}" );
 
 		// Get a list of 5 Companies
 		if( companies.TotalCount < 5 )
 		{
-			_ = Program.sLogger.Info( "Companies count is less than 5!" );
+			_ = Program.sLogger.Info( "Companies total is less than 5!" );
 			return false;
 		}
 		IList<ICompany> listc = companies.Get( 5 );
@@ -53,23 +58,40 @@ internal class TestAPI
 
 		#region Test the Person data
 
-		People people = new( Program.sApp.ConfigFile );
+		People people = new( ctx );
 		_ = Program.sLogger.Info( $"People total....: {people.TotalCount:00#}" );
 
 		// Get a list of 10 People
 		if( people.TotalCount < 10 )
 		{
-			_ = Program.sLogger.Info( "People count is less than 10!" );
+			_ = Program.sLogger.Info( "People total is less than 10!" );
 			return false;
 		}
 		IList<IPerson> listp = people.Get( 10 );
 		_ = Program.sLogger.Info( $"People list.....: {listp.Count:00#}" );
 
 		// Get a specific Person
-		IPerson? person = people.Find( listc[0].Id );
+		IPerson? person = people.Find( listp[0].Id );
 		_ = person is not null
 			? Program.sLogger.Info( $"Person Id  {listp[0].Id:00#} is {person.FullName}" )
 			: Program.sLogger.Info( $"Person Id {listp[0].Id:00#} not found!" );
+
+		#endregion
+
+		#region Test the Additional data
+
+		Console.WriteLine();
+		_ = Program.sLogger.Info( $"Movies total....: {ctx.Movies.Count():00#}" );
+
+		// Get a specific Movie
+		Common.Models.Movie? movie = ctx.Movies.Find( [5] );
+		_ = Program.sLogger.Info( $"Movie title.....: {movie?.Title}" );
+
+		_ = Program.sLogger.Info( $"SuperHero total.: {ctx.SuperHeroes.Count():00#}" );
+
+		// Get a specific SuperHero
+		Common.Models.SuperHero? hero = ctx.SuperHeroes.Find( [35] );
+		_ = Program.sLogger.Info( $"SuperHero name..: {hero?.Name}" );
 
 		#endregion
